Move responses to better suited package
